added task 3

+++ math/0x03-probability/exponential.py
@@ -19,4 +19,4 @@ class Exponential:
                 raise TypeError("data must be a list")
             if len(data) < 2:
                 raise ValueError("data must contain multiple values")
-            self.lambtha = 1 / (sum(data) / len(data))
+            self.lambtha = float(1 / (sum(data) / len(data)))
